pyproject.toml
[tool.poetry]
name = "clip_service"
version = "0.1.0"
description = ""
authors = [

]
maintainers = [

]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.85.0"
uvicorn = { version = "^0.18.3", extras = ["standard"] }
pydantic = {version = "^1.10.2", extras = ["dotenv"]}
yarl = "^1.8.1"
ujson = "^5.5.0"
aiofiles = "^22.1.0"
httptools = "^0.5.0"
loguru = "^0.6.0"
torch = { version = "1.11.0+cpu", source = "torchcpu" }
onnxruntime = "1.11.1"
onnx = "1.11.0"
torchvision = "0.12.0+cpu"
Pillow = "^9.3.0"

[[tool.poetry.source]]
name = "torchcpu"
url = "https://download.pytorch.org/whl/cpu"
default = false
secondary = true

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"
flake8 = "~4.0.1"
mypy = "^0.981"
isort = "^5.10.1"
yesqa = "^1.4.0"
pre-commit = "^2.20.0"
wemake-python-styleguide = "^0.17.0"
black = "^22.8.0"
autoflake = "^1.6.1"
pytest-cov = "^4.0.0"
anyio = "^3.6.1"
pytest-env = "^0.6.2"
httpx = "^0.23.0"

[tool.isort]
profile = "black"
multi_line_output = 3
src_paths = ["clip_service",]

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true
warn_unused_ignores = false
warn_return_any = false
namespace_packages = true

[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore:.*unclosed.*:ResourceWarning",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
